Consolidate components into feature folders
