+++ src/features/filters/FilterLink.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
-import { setVisibilityFilter } from 'features/filters/filtersSlice'
-import Link from '../components/Link'
+import { setVisibilityFilter } from './filtersSlice'
+import Link from './Link'
 
 const mapStateToProps = (state, ownProps) => ({
   active: ownProps.filter === state.visibilityFilter
